pyproject.toml
[project]
name = "prospect-ai"
version = "0.1.0"
description = "AI Sales Outreach Multi-Agent System"
requires-python = ">=3.11,<3.14"

dependencies = [
    # Web server
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Google OAuth2 + session
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.1.0",
    # Agent framework
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    # LLM
    "openai>=1.0.0",
    # Data handling
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    # Browser automation
    "playwright>=1.44.0",
    # MCP client (Playwright MCP connection)
    "mcp>=1.0.0",
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.3.0",
]
full = [
    "browser-use>=0.1.0",
    "resend>=0.7.0",
    "sendgrid>=6.0.0",
    "sqlalchemy>=2.0.0",
]

[tool.uv]
python-preference = "managed"
